split into two artifacts - jakarta and jaxb #70
