+++ CHANGELOG.md
@@ -1,46 +1,46 @@
 # Changelog
 
-## [0.2.0](https://github.com/sokomishalov/jackson-dataformat-soap/tree/0.2.0) (2022-04-17)
+## [0.2.0](https://github.com/sokomishalov/jackson-datatype-soap/tree/0.2.0) (2022-04-17)
 
-[Full Changelog](https://github.com/sokomishalov/jackson-dataformat-soap/compare/0.1.1...0.2.0)
+[Full Changelog](https://github.com/sokomishalov/jackson-datatype-soap/compare/0.1.1...0.2.0)
 
 **Closed issues:**
 
-- does not run with jackson-module-jaxb-annotations 2.13. [\#37](https://github.com/sokomishalov/jackson-dataformat-soap/issues/37)
-- Support multiple header classes [\#36](https://github.com/sokomishalov/jackson-dataformat-soap/issues/36)
+- does not run with jackson-module-jaxb-annotations 2.13. [\#37](https://github.com/sokomishalov/jackson-datatype-soap/issues/37)
+- Support multiple header classes [\#36](https://github.com/sokomishalov/jackson-datatype-soap/issues/36)
 
 **Merged pull requests:**
 
-- fix \#36 [\#40](https://github.com/sokomishalov/jackson-dataformat-soap/pull/40) ([sokomishalov](https://github.com/sokomishalov))
-- Fix \#37 [\#39](https://github.com/sokomishalov/jackson-dataformat-soap/pull/39) ([sokomishalov](https://github.com/sokomishalov))
-- Bump dokka-maven-plugin from 1.5.0 to 1.5.30 [\#16](https://github.com/sokomishalov/jackson-dataformat-soap/pull/16) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump maven-javadoc-plugin from 3.3.0 to 3.3.1 [\#15](https://github.com/sokomishalov/jackson-dataformat-soap/pull/15) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump jackson-bom from 2.12.4 to 2.12.5 [\#14](https://github.com/sokomishalov/jackson-dataformat-soap/pull/14) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump kotlin.version from 1.5.21 to 1.5.30 [\#13](https://github.com/sokomishalov/jackson-dataformat-soap/pull/13) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump kotlin.version from 1.5.20 to 1.5.21 [\#12](https://github.com/sokomishalov/jackson-dataformat-soap/pull/12) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump dokka-maven-plugin from 1.4.32 to 1.5.0 [\#11](https://github.com/sokomishalov/jackson-dataformat-soap/pull/11) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump jackson-bom from 2.12.3 to 2.12.4 [\#10](https://github.com/sokomishalov/jackson-dataformat-soap/pull/10) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump kotlin.version from 1.5.10 to 1.5.20 [\#9](https://github.com/sokomishalov/jackson-dataformat-soap/pull/9) ([dependabot[bot]](https://github.com/apps/dependabot))
+- fix \#36 [\#40](https://github.com/sokomishalov/jackson-datatype-soap/pull/40) ([sokomishalov](https://github.com/sokomishalov))
+- Fix \#37 [\#39](https://github.com/sokomishalov/jackson-datatype-soap/pull/39) ([sokomishalov](https://github.com/sokomishalov))
+- Bump dokka-maven-plugin from 1.5.0 to 1.5.30 [\#16](https://github.com/sokomishalov/jackson-datatype-soap/pull/16) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump maven-javadoc-plugin from 3.3.0 to 3.3.1 [\#15](https://github.com/sokomishalov/jackson-datatype-soap/pull/15) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump jackson-bom from 2.12.4 to 2.12.5 [\#14](https://github.com/sokomishalov/jackson-datatype-soap/pull/14) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump kotlin.version from 1.5.21 to 1.5.30 [\#13](https://github.com/sokomishalov/jackson-datatype-soap/pull/13) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump kotlin.version from 1.5.20 to 1.5.21 [\#12](https://github.com/sokomishalov/jackson-datatype-soap/pull/12) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump dokka-maven-plugin from 1.4.32 to 1.5.0 [\#11](https://github.com/sokomishalov/jackson-datatype-soap/pull/11) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump jackson-bom from 2.12.3 to 2.12.4 [\#10](https://github.com/sokomishalov/jackson-datatype-soap/pull/10) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump kotlin.version from 1.5.10 to 1.5.20 [\#9](https://github.com/sokomishalov/jackson-datatype-soap/pull/9) ([dependabot[bot]](https://github.com/apps/dependabot))
 
-## [0.1.1](https://github.com/sokomishalov/jackson-dataformat-soap/tree/0.1.1) (2021-06-21)
+## [0.1.1](https://github.com/sokomishalov/jackson-datatype-soap/tree/0.1.1) (2021-06-21)
 
-[Full Changelog](https://github.com/sokomishalov/jackson-dataformat-soap/compare/0.1.0...0.1.1)
-
-**Merged pull requests:**
-
-- remove collection setters [\#8](https://github.com/sokomishalov/jackson-dataformat-soap/pull/8) ([sokomishalov](https://github.com/sokomishalov))
-- Bump maven-gpg-plugin from 1.6 to 3.0.1 [\#4](https://github.com/sokomishalov/jackson-dataformat-soap/pull/4) ([dependabot[bot]](https://github.com/apps/dependabot))
-
-## [0.1.0](https://github.com/sokomishalov/jackson-dataformat-soap/tree/0.1.0) (2021-06-09)
-
-[Full Changelog](https://github.com/sokomishalov/jackson-dataformat-soap/compare/e943b1bc112c28c6c04f09ccecbfc6e8d20f294b...0.1.0)
+[Full Changelog](https://github.com/sokomishalov/jackson-datatype-soap/compare/0.1.0...0.1.1)
 
 **Merged pull requests:**
 
-- Bump cxf-codegen-plugin from 3.4.3 to 3.4.4 [\#7](https://github.com/sokomishalov/jackson-dataformat-soap/pull/7) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump kotlin.version from 1.4.32 to 1.5.10 [\#6](https://github.com/sokomishalov/jackson-dataformat-soap/pull/6) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump maven-javadoc-plugin from 3.2.0 to 3.3.0 [\#5](https://github.com/sokomishalov/jackson-dataformat-soap/pull/5) ([dependabot[bot]](https://github.com/apps/dependabot))
-- Bump cxf-codegen-plugin from 3.3.7 to 3.4.3 [\#3](https://github.com/sokomishalov/jackson-dataformat-soap/pull/3) ([dependabot[bot]](https://github.com/apps/dependabot))
+- remove collection setters [\#8](https://github.com/sokomishalov/jackson-datatype-soap/pull/8) ([sokomishalov](https://github.com/sokomishalov))
+- Bump maven-gpg-plugin from 1.6 to 3.0.1 [\#4](https://github.com/sokomishalov/jackson-datatype-soap/pull/4) ([dependabot[bot]](https://github.com/apps/dependabot))
+
+## [0.1.0](https://github.com/sokomishalov/jackson-datatype-soap/tree/0.1.0) (2021-06-09)
+
+[Full Changelog](https://github.com/sokomishalov/jackson-datatype-soap/compare/e943b1bc112c28c6c04f09ccecbfc6e8d20f294b...0.1.0)
+
+**Merged pull requests:**
+
+- Bump cxf-codegen-plugin from 3.4.3 to 3.4.4 [\#7](https://github.com/sokomishalov/jackson-datatype-soap/pull/7) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump kotlin.version from 1.4.32 to 1.5.10 [\#6](https://github.com/sokomishalov/jackson-datatype-soap/pull/6) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump maven-javadoc-plugin from 3.2.0 to 3.3.0 [\#5](https://github.com/sokomishalov/jackson-datatype-soap/pull/5) ([dependabot[bot]](https://github.com/apps/dependabot))
+- Bump cxf-codegen-plugin from 3.3.7 to 3.4.3 [\#3](https://github.com/sokomishalov/jackson-datatype-soap/pull/3) ([dependabot[bot]](https://github.com/apps/dependabot))
