+++ README.md
@@ -1,9 +1,10 @@
-jackson-dataformat-soap
+jackson-datatype-soap
 ========
+~~Here should be some fancy logo~~
 
 [![Apache License 2](https://img.shields.io/badge/license-ASF2-blue.svg)](https://choosealicense.com/licenses/apache-2.0/)
-[![](https://img.shields.io/maven-central/v/ru.sokomishalov.jackson/jackson-dataformat-soap)](https://mvnrepository.com/artifact/ru.sokomishalov.jackson/jackson-dataformat-soap)
-[![](https://img.shields.io/jitpack/v/github/sokomishalov/jackson-dataformat-soap)](https://jitpack.io/#sokomishalov/jackson-dataformat-soap)
+[![](https://img.shields.io/maven-central/v/ru.sokomishalov.jackson/jackson-datatype-soap)](https://mvnrepository.com/artifact/ru.sokomishalov.jackson/jackson-datatype-soap-jakarta)
+[![](https://img.shields.io/jitpack/v/github/sokomishalov/jackson-datatype-soap)](https://jitpack.io/#sokomishalov/jackson-datatype-soap)
 
 ## Overview
 
@@ -14,10 +15,18 @@ Jackson SOAP implementation over jackson-dataformat-xml
 Maven:
 
 ```xml
-
+<!-- Legacy (JAXB) artifact -->
 <dependency>
     <groupId>ru.sokomishalov.jackson</groupId>
-    <artifactId>jackson-dataformat-soap</artifactId>
+    <artifactId>jackson-datatype-soap-jaxb</artifactId>
+    <version>x.y.z</version>
+</dependency>
+```
+```xml
+<!-- New (Jakarta) artifact -->
+<dependency>
+    <groupId>ru.sokomishalov.jackson</groupId>
+    <artifactId>jackson-datatype-soap-jakarta</artifactId>
     <version>x.y.z</version>
 </dependency>
 ```
@@ -25,7 +34,10 @@ Maven:
 Gradle kotlin dsl:
 
 ```kotlin
-implementation("ru.sokomishalov.jackson:jackson-dataformat-soap:x.y.z")
+// Legacy (JAXB) artifact
+implementation("ru.sokomishalov.jackson:jackson-datatype-soap-jaxb:x.y.z")
+// New (Jakarta) artifact
+implementation("ru.sokomishalov.jackson:jackson-datatype-soap-jakarta:x.y.z")
 ```
 
 ## Usage
