+++ jackson-datatype-soap-jakarta/src/test/resources/example/soap_fault_1_1.xml
