+++ jackson-datatype-soap-jakarta/src/test/resources/wsdl/issue-36.wsdl
