+++ jackson-datatype-soap-jaxb/src/main/resources/META-INF/services/com.fasterxml.jackson.databind.ObjectMapper
@@ -1,0 +1,1 @@
+ru.sokomishalov.jackson.datatype.soap.SoapMapper
