+++ jackson-datatype-soap-jaxb/src/test/resources/example/soap_fault_1_1.xml
@@ -1,0 +1,9 @@
+<soap:Envelope xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/">
+    <soap:Body>
+        <soap:Fault>
+            <faultcode>soap:VersionMismatch</faultcode>
+            <faultstring>Message was not SOAP 1.1 compliant</faultstring>
+            <faultactor>http://sample.org.ocm/jws/authnticator</faultactor>
+        </soap:Fault>
+    </soap:Body>
+</soap:Envelope>
