+++ jackson-datatype-soap-jaxb/src/test/resources/example/soap_fault_1_2.xml
@@ -1,0 +1,15 @@
+<soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope">
+    <soap:Body>
+        <soap:Fault>
+            <soap:Code>
+                <soap:Value>env:Sender</soap:Value>
+                <soap:Subcode>
+                    <soap:Value>rpc:BadArguments</soap:Value>
+                </soap:Subcode>
+            </soap:Code>
+            <soap:Reason>
+                <soap:Text xml:lang="en-US">Processing error</soap:Text>
+            </soap:Reason>
+        </soap:Fault>
+    </soap:Body>
+</soap:Envelope>
