--- src/main/resources/META-INF/services/com.fasterxml.jackson.databind.ObjectMapper
@@ -1,1 +1,0 @@
-ru.sokomishalov.jackson.dataformat.soap.SoapMapper
